modified legalp
Now, legalp(parameters) accpet a list as its argument. The first member of the list
is bet, the second is choice, the third the current score of this agent.
By using a list as one parameter rather than three parameters, we have
the benefit that we can check the number of inputs and when the number
of inputs is invalid it treats this as an invalid input, rather than
throws error and haults the whole program.